Coursera Capstone Project Main Commit
In this project you are required to create a data model for Little Lemon and deploy it in MySQL database. This will include a set of tables to store relevant information. Once you have completed this task, feel free to populate your tables with some records of you own data for testing purposes.

In addition you also need to analyze data using the following Little Lemon data file:Once you have created the appropriate models, database schema, and completed the data analysis in Tableau, you then need to commit the project to your GitHub and upload the link.

Your project is expected to include the following items:

Relevant diagrams and screenshots in PNG format.

The data model you produced in MySQL Workbench.

The SQL file of the Little Lemon database schema.

The workbook that includes all worksheets and dashboards.